Move normalized_column_identier test to sql_integ
